check box
Longer term, I would like to see that we make entering to sweepstakes
optional with check box on the form that will be checked by default.
And the actual Title/copy closer to something like this:

<Doctor ID> inviting you to watch a brief video on the importance of
preserving your baby's cord blood and tissue at birth.
This video was developed in collaboration with Parent's Guide to Cord
Blood Foundation to educate future parents about life saving options.
Please fill in the form below to proceed :

There the <Doctor ID> is the value we enter in the iPad configuration.
(The name of practice or a doctors name)
I would like this to sound more authoritative (Doctor invites) and
independent PGCB. I would expect much higher wiliness fill in the form
then just for the purpose of entering to sweepstakes.